@@ -1,4 +1,6 @@
-const nextConfig = {
+import type { NextConfig } from "next";
+
+const nextConfig: NextConfig = {
   output: "export", // Keep this if you're using static export
   images: {
     unoptimized: true, // Remove if you want Next.js to optimize images on Vercel
@@ -6,4 +8,4 @@ const nextConfig = {
   trailingSlash: false, // Vercel does not need trailing slashes
 };
 
-module.exports = nextConfig;
+export default nextConfig;
